+++ packages/express/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 2.1.61
+
+### Patch Changes
+
+- Updated dependencies [[`ea8cb05`](https://github.com/clerk/javascript/commit/ea8cb055cecd986425f75b2f2da9cfec8a4b2ff4), [`f541dbf`](https://github.com/clerk/javascript/commit/f541dbfeea2b840409733222d40b53b0580e7c47)]:
+  - @clerk/shared@4.29.3
+  - @clerk/backend@3.16.10
+
 ## 2.1.60
 
 ### Patch Changes