+++ packages/hono/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @clerk/hono
 
+## 0.1.71
+
+### Patch Changes
+
+- Updated dependencies [[`ea8cb05`](https://github.com/clerk/javascript/commit/ea8cb055cecd986425f75b2f2da9cfec8a4b2ff4), [`f541dbf`](https://github.com/clerk/javascript/commit/f541dbfeea2b840409733222d40b53b0580e7c47)]:
+  - @clerk/shared@4.29.3
+  - @clerk/backend@3.16.10
+
 ## 0.1.70
 
 ### Patch Changes