+++ packages/tanstack-react-start/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @clerk/tanstack-react-start
 
+## 1.5.6
+
+### Patch Changes
+
+- Updated dependencies [[`ea8cb05`](https://github.com/clerk/javascript/commit/ea8cb055cecd986425f75b2f2da9cfec8a4b2ff4), [`f541dbf`](https://github.com/clerk/javascript/commit/f541dbfeea2b840409733222d40b53b0580e7c47)]:
+  - @clerk/shared@4.29.3
+  - @clerk/backend@3.16.10
+  - @clerk/react@6.14.5
+
 ## 1.5.5
 
 ### Patch Changes